Created a clone of Flappy Bird with simple rectangles.
This pygame imitates a simple version of FLappy Bird. The Bird is acted on by gravity and the flapping input of the player.
Lose conditions are going out of bounds either up or down, or hitting a pipe.
You gain points by clearing the gap between pipes.
Pipes spawan regularly with a random position for the gap.

	new file:   GLOBAL.py
	new file:   bird.py
	new file:   main.py
	new file:   pipe.py